fix: Speedup resize, part 1
