+++ apps/builder/app/builder/features/workspace/index.ts
@@ -1,3 +1,2 @@
 export * from "./workspace";
-export * from "./use-read-canvas-rect";
 export * from "./canvas-iframe";
